do not expose existing StripeSync, just the new one
